fixed compilation problems, merged BFS and DFS into Graphs class
