+++ java/BinarySearch.java
@@ -1,5 +1,5 @@
 
-public class Main {
+public class BinarySearch {
 
     public static void main(String[] args) {
         int[] arr = {1, 2, 3, 4, 5, 6, 7, 8, 9, 10};
