+++ java/Graph.java
@@ -1,22 +1,41 @@
 import java.io.*;
 import java.util.*;
 
-class Graph
-{
-    private int V;   // No. of vertices
-    private LinkedList<Integer> adj[]; //Adjacency Lists
-
-    Graph(int v)
+class Graph {
+    private int V; 
+    private LinkedList<Integer> adj[];
+  
+    @SuppressWarnings("unchecked") Graph(int v)
     {
         V = v;
         adj = new LinkedList[v];
-        for (int i=0; i<v; ++i)
+        for (int i = 0; i < v; ++i)
             adj[i] = new LinkedList();
     }
-
-    void addEdge(int v,int w)
+  
+    void addEdge(int v, int w)
     {
-        adj[v].add(w);
+        adj[v].add(w); // Add w to v's list.
+    }
+  
+    void DFSUtil(int v, boolean visited[])
+    {
+        visited[v] = true;
+        System.out.print(v + " ");
+  
+        Iterator<Integer> i = adj[v].listIterator();
+        while (i.hasNext()) 
+        {
+            int n = i.next();
+            if (!visited[n])
+                DFSUtil(n, visited);
+        }
+    }
+  
+    void DFS(int v)
+    {
+        boolean visited[] = new boolean[V];
+        DFSUtil(v, visited);
     }
 
     void BFS(int s)
@@ -25,10 +44,10 @@ class Graph
         boolean visited[] = new boolean[V];
 
         LinkedList<Integer> queue = new LinkedList<Integer>();
- 
+
         visited[s]=true;
         queue.add(s);
- 
+
         while (queue.size() != 0)
         {
             s = queue.poll();
@@ -46,21 +65,24 @@ class Graph
             }
         }
     }
-
+  
+ 
     public static void main(String args[])
     {
         Graph g = new Graph(4);
- 
+  
         g.addEdge(0, 1);
         g.addEdge(0, 2);
         g.addEdge(1, 2);
         g.addEdge(2, 0);
         g.addEdge(2, 3);
         g.addEdge(3, 3);
- 
-        System.out.println("Following is Breadth First Traversal "+
-                           "(starting from vertex 2)");
- 
+  
+        System.out.println(
+            "Following is Depth First Traversal "
+            + "(starting from vertex 2)");
+  
+        g.DFS(2);
         g.BFS(2);
     }
 }
