+++ java/PalindromeNumber.java
@@ -1,6 +1,5 @@
 import java.util.*;
-public class Main
-{
+public class PalindromeNumber {
 	public static void main(String[] args) {
 		Scanner scan=new Scanner(System.in);
 		int n=scan.nextInt();
