+++ java/factorial.java
@@ -1,6 +1,6 @@
 import java.util.*;
 
-public class Main {
+public class factorial {
     public static void main(String[] args) {
         Scanner scan = new Scanner(System.in);
         System.out.println("Enter the element");
